Moving bomb files to Square pkg
