+++ core/src/main/java/com/bomber7/core/model/square/TimeBomb.java
@@ -1,4 +1,4 @@
-package src.main.java.com.bomber7.core.model;
+package com.bomber7.core.model.square;
 
 public class TimeBomb extends Bomb {
     private int timer;
